Split the Applications page into smaller chunks
